# docs/open-api-spec.json
{
    "components": {
    },
    "info": {
        "contact": {
            "email": "s.sosnik@gmail.com",
            "name": "Sebastian Sosnik",
            "url": "https://github.com/ssosnik/app-rating"
        },
        "title": "Apps Rating OpenAPI definition",
        "description": "Filter and sort apps by rating",
        "version": "v1.0"
    },
    "openapi": "3.0.1",
    "paths": {
        "/api/{appUuid}/avg": {
            "get": {
                "operationId": "appAverageRating",
                "parameters": [
                    {
                        "in": "path",
                        "name": "appUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "since",
                        "required": true,
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "until",
                        "required": true,
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "app-rating-controller"
                ]
            }
        },
        "/api/top-apps/{ageGroup}": {
            "get": {
                "operationId": "topAppsForAgeGroup",
                "parameters": [
                    {
                        "in": "path",
                        "name": "ageGroup",
                        "required": true,
                        "schema": {
                            "enum": [
                                "AGE_GROUP_1",
                                "AGE_GROUP_2",
                                "AGE_GROUP_3",
                                "AGE_GROUP_4",
                                "AGE_GROUP_5",
                                "AGE_GROUP_6",
                                "AGE_GROUP_7"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "since",
                        "required": true,
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "until",
                        "required": true,
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "app-rating-controller"
                ]
            }
        }
    },
    "servers": [
        {
            "description": "Generated server url",
            "url": "http://localhost:8080"
        }
    ]
}